# docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /answer/api/v1/chat/completion:
    get:
      consumes:
      - application/json
      description: Get Chat Completion
      parameters:
      - description: Prompt for the chat completion
        in: query
        name: prompt
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Chat Completion
      tags:
      - api-answer
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"